Add typeahead for unix console input
At the console, editline will check for the cursor position; on unix,
this is done by writing an ansi command sequence and then reading the
result. Keyboard might happen before the result. The fix is to add
typeahead to buffer the keyboard input while reading for the command
result.

Fixes #29